+++ crates/burn-import/onnx-tests/tests/argmax/mod.rs
@@ -1,0 +1,25 @@
+// Import the shared macro
+use crate::include_models;
+include_models!(argmax);
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use burn::tensor::{Tensor, TensorData};
+
+    type Backend = burn_ndarray::NdArray<f32>;
+
+    #[test]
+    fn argmax() {
+        // Initialize the model with weights (loaded from the exported file)
+        let model: argmax::Model<Backend> = argmax::Model::default();
+
+        let device = Default::default();
+        // Run the model
+        let input = Tensor::<Backend, 2>::from_floats([[1., 2., 3.], [4., 5., 6.]], &device);
+        let output = model.forward(input);
+        let expected = TensorData::from([[2i64], [2]]);
+
+        output.to_data().assert_eq(&expected, true);
+    }
+}
