+++ crates/burn-import/onnx-tests/tests/test_mod.rs
@@ -7,6 +7,7 @@ mod node_tests;
 
 // Import individual node modules
 pub mod add;
+pub mod argmax;
 pub mod concat;
 pub mod constant;
 pub mod constant_of_shape;
@@ -16,8 +17,15 @@ pub mod div;
 pub mod dropout;
 pub mod erf;
 pub mod gather;
+pub mod global_avr_pool;
+pub mod log_softmax;
 pub mod matmul;
+pub mod max;
 pub mod mean;
+pub mod min;
 pub mod mul;
+pub mod slice;
+pub mod softmax;
+pub mod sqrt;
 pub mod sub;
 pub mod sum;
